refactor: relocate findRelevantSession to user utility
